requirements.txt
pandas==1.1.5
dateparser==1.1.0
python-dotenv==0.19.2
web3==5.25.0
gql==3.0.0rc0